power: qpnp-fg: make CHARGE_FULL property writable by user
CHARGE_FULL property indicates the learnt capacity of the battery
from the last capacity learning cycle. However, in some rare
cases, if the learnt capacity is having an unexpected value, it
requires a reset to a good value which is not possible for an
user who cannot reinsert a battery or do an explicit write to FG
SRAM. Allowing CHARGE_FULL to be writable helps this.

Bug:78883741

Change-Id: Ibf4760505e345f504be4d5372b1e41558ea6fb86
Signed-off-by: Jack Wu <wjack@google.com>
Signed-off-by: utsavbalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: utsavisgreat <utsavbalar1231@gmail.com>
Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>
Signed-off-by: Ahmad Nauval Syahputra <anvalputra@gmail.com>