Makefile: Rework on Polly optimization Flags
Some flags like:
1. -polly-reschedule
2. -polly-loopfusion-greedy
3. -polly-postopts

were introduced only in Clang 14, Which means when a user tries to
compile the kernel with Clang 13 or lower, it will error out.

Retain compatibility for older clangs by passing the above given flags only if kernel is compiled with clang 14,
If not, pass '-polly-opt-fusion' which is a flag that got removed in clang 14.

Also, pass polly flags to KBUILD_AFLAGS and KBUILD_LDFLAGS as well.

Signed-off-by: Ahmad Nauval Syahputra <anvalputra@gmail.com>